refactor: implemented USB CDC communication

+++ devices/raspberry_pi_pico_2w/tinygo/internal/escmotor/constants.go
@@ -1,7 +1,7 @@
 package escmotor
 
 const (
-	DefaultIsPolarityInverted      = false
-	DefaultMinPulse           uint = 1000
-	DefaultMaxPulse           uint = 2000
+	DefaultIsPolarityInverted        = false
+	DefaultMinPulse           uint16 = 1000
+	DefaultMaxPulse           uint16 = 2000
 )
